+++ ivy_tests/test_ivy/test_frontends/test_jax/test_jax_numpy_manipuration.py
@@ -2,6 +2,7 @@
 from hypothesis import strategies as st
 
 # local
+import ivy
 import ivy_tests.test_ivy.helpers as helpers
 from ivy_tests.test_ivy.helpers import handle_frontend_test
 import ivy_tests.test_ivy.test_frontends.test_numpy.helpers as np_frontend_helpers
@@ -536,4 +537,123 @@ def test_jax_numpy_take(
         a=value,
         indices=indices,
         axis=axis,
+    )
+
+
+# uint16
+@handle_frontend_test(
+    fn_tree="jax.numpy.uint16",
+    dtype_and_x=helpers.dtype_and_values(
+        available_dtypes=helpers.get_dtypes("integer"),
+    ),
+)
+def test_jax_numpy_uint16(
+    *,
+    dtype_and_x,
+    num_positional_args,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    input_dtype, x = dtype_and_x
+    if ivy.current_backend_str() != "torch":
+        helpers.test_frontend_function(
+            input_dtypes=input_dtype,
+            as_variable_flags=as_variable,
+            with_out=False,
+            num_positional_args=num_positional_args,
+            native_array_flags=native_array,
+            frontend=frontend,
+            fn_tree=fn_tree,
+            on_device=on_device,
+            x=x[0],
+        )
+
+
+# asarray
+@handle_frontend_test(
+    fn_tree="jax.numpy.asarray",
+    dtype_and_a=helpers.dtype_and_values(
+        available_dtypes=helpers.get_dtypes("numeric"),
+        num_arrays=1,
+        min_num_dims=0,
+        max_num_dims=5,
+        min_dim_size=1,
+        max_dim_size=5,
+    ),
+)
+def test_jax_numpy_asarray(
+    dtype_and_a,
+    as_variable,
+    num_positional_args,
+    native_array,
+    frontend,
+    fn_tree,
+    on_device,
+):
+    dtype, a = dtype_and_a
+    helpers.test_frontend_function(
+        input_dtypes=dtype,
+        as_variable_flags=as_variable,
+        with_out=False,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        a=a,
+        dtype=dtype[0],
+    )
+
+
+# broadcast_to
+@st.composite
+def _get_input_and_broadcast_shape(draw):
+    dim1 = draw(helpers.ints(min_value=2, max_value=5))
+    x_dtype, x = draw(
+        helpers.dtype_and_values(
+            available_dtypes=helpers.get_dtypes("valid"),
+            min_num_dims=1,
+            max_num_dims=5,
+            min_dim_size=2,
+            max_dim_size=10,
+            shape=(dim1,),
+        )
+    )
+    broadcast_dim = draw(helpers.ints(min_value=1, max_value=3))
+    shape = ()
+    for _ in range(broadcast_dim):
+        shape += (draw(helpers.ints(min_value=1, max_value=dim1)),)
+    shape += (dim1,)
+    return x_dtype, x, shape
+
+
+@handle_frontend_test(
+    fn_tree="jax.numpy.broadcast_to",
+    input_x_broadcast=_get_input_and_broadcast_shape(),
+)
+def test_jax_numpy_broadcast_to(
+    *,
+    input_x_broadcast,
+    num_positional_args,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    x_dtype, x, shape = input_x_broadcast
+    helpers.test_frontend_function(
+        input_dtypes=x_dtype,
+        as_variable_flags=as_variable,
+        with_out=False,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        arr=x[0],
+        shape=shape,
     )
